Tasks/secure-nginx-cert-rotation/requirements.txt
pytest==8.2.2
requests==2.32.3
pyopenssl==24.1.0